# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Keep in touch with your friends by sharing photos of
    special moments, thanks to WASAPhoto! You can upload your photos
    directly from your PC, and they will be visible to everyone
    following you.
  version: 1.0.0
...
paths:
...
  /users/{uid}/stream:
    parameters:
    - $ref: "#/components/parameters/userid"

    get:
      tags: ["stream"]
      summary: returns the stream of the user
      description: |
        the stream of the user, containing the photos published
        by all the users followed in reverse chronological order
      operationId: getMyStream
      responses:
        "200":
          description: the stream is returned in the content
          content:
            application/json:
              schema:
                description: |
                  an array containing the photos published by the users
                  followed by the user, in reverse chronological order
                type: array
                minItems: 0
                maxItems: 99999
                items:
                  $ref: "#/components/schemas/photo"
        "401":
          description: the user has no right to access this resource
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
...
  /users/{uid}/photos/{photo}:
    parameters:
    - $ref: "#/components/parameters/userid"
    - $ref: "#/components/parameters/photoParam"

    get:
      tags: ["photos"]
      summary: returns a photo
      description: a user can access a photo if e/she is the owner or if the owner has not banned him/her
      operationId: getPhoto
      responses:
        "200":
          description: the photo is returned in the content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/photo"
        "401":
          description: the user has no right to access this resource
        "403":
          description: the owner of the photo has banned the user that called the operation
        "404":
          description: photo not found

      security:
        - bearerAuth: []

    delete:
      tags: ["photos"]
      summary: remove a photo from the user's photos
      description: |
        the photo is deleted and won't appear anymore
        in the user's photos
      operationId: deletePhoto
      responses:
        "204":
          description: the photo is successfully removed from the array of photos
        "401":
          description: the user has no right to access this resource
        "403":
          description: can't delete a photo published by someone else
        "404":
          description: can't find the photo
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
  /users/{uid}/photos/{photo}/likes:
    parameters:
    - $ref: "#/components/parameters/userid"
    - $ref: "#/components/parameters/photoParam"

    get:
      tags: ["likes"]
      summary: the likes on the photo
      description: |
        this operation returns the number of likes and the users that like the photo
      operationId: getLikeOnPhoto
      responses:
        "200":
          description: |
            the number of likes and the users that like
            the photo are returned in the content
          content:
            application/json:
              schema:
                description: |
                  an object representing the like counter
                  and the array containing the partial profile of the users
                  that like the photo
                type: object
                properties:
                  likeCount:
                    description: how many users like the photo
                    type: integer
                    minimum: 0
                    maximum: 99999
                    example: 10
                  likingUsers:
                    description: |
                      an array containing the partial profile of the users
                      that like the photo
                    type: array
                    minItems: 0
                    maxItems: 99999
                    items:
                      $ref: "#/components/schemas/partialProfile"
        "401":
          description: the user has no right to access this resource
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
...
  /users/{uid}/photos/{photo}/comments:
    parameters:
    - $ref: "#/components/parameters/userid"
    - $ref: "#/components/parameters/photoParam"

    get:
      tags: ["comments"]
      summary: the comments under a photo
      description: |
        the operation returns the number of comments on the photo and
        all the comments published on it
      operationId: getComments
      responses:
        "200":
          description: |
            the comment counter and the comments published
            on the photo are listed in the content
          content:
            application/json:
              schema:
                description: |
                  an object representing the comment counter
                  and the array containing the comments
                type: object
                properties:
                  commentCount:
                    description: how many comments were left on the photo
                    type: integer
                    minimum: 0
                    maximum: 99999
                    example: 3
                  commentArray:
                    description: an array containing the comments on the photo
                    type: array
                    minItems: 0
                    maxItems: 99999
                    items:
                      $ref: "#/components/schemas/comment"
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []
    
    post:
      tags: ["comments"]
      summary: comment a photo
      description: leave a comment on a photo
      operationId: commentPhoto
      requestBody:
        description: the text that uid want to post as a comment
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/commentOnReq"
        required: true
      responses:
        "201":
          description: |
            the comment is added to the comment array of the photo
            and the comment counter is updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/comment"
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []



  #####################################################################
...
  /users/{uid}/followers:
    parameters:
    - $ref: "#/components/parameters/userid"

    get:
      tags: ["followers"]
      summary: list all the users that are following the user
      description: |
        the operation returns a list of all the users
        that are following the user
      operationId: getFollowers
      responses:
        "200":
          description: |
            the partial profile of the users that are currently following this user
            are listed in the content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/followers"
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
  /users/{uid}/following:
    parameters:
    - $ref: "#/components/parameters/userid"

    get:
      tags: ["following"]
      summary: list all the users followed
      description: |
        returns the list of the uid of all the users currently followed
      operationId: getFollowing
      responses:
        "200":
          description: |
            the partial profile of the users that are currently followed by this user
            are listed in the content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/following"
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
...
  /users/{uid}/banned:
    parameters:
    - $ref: "#/components/parameters/userid"

    get:
      tags: ["ban"]
      summary: list all the users currently banned
      description: |
        returns the list of the uid of all the users currently banned
      operationId: getBanned
      responses:
        "200":
          description: |
            the user id of all the users banned by the user
            are listed in the content
          content:
            application/json:
              schema:
                description: |
                  an array of strings, where every string is the partial profile
                  of one of the users banned
                type: array
                minItems: 0
                maxItems: 99999
                items:
                  $ref: "#/components/schemas/partialProfile"
        "401":
          description: the user has no right to access this resource
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
  /users/{uid}/banned/{bannedId}:
    parameters:
    - $ref: "#/components/parameters/userid"
    - $ref: '#/components/parameters/bannedUserId'

    put:
      tags: ["ban"]
      summary: ban another user
      description: bannedId is added to the user's banned list
      operationId: banUser
      responses:
        "204":
          description: bannedId is added to the user's banned list
        "401":
          description: the user has no right to access this resource
        "403":
          description: the user is in bannedId's banned list
        "404":
          description: can't find bannedId
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

    delete:
      tags: ["ban"]
      summary: unban a user
      description: bannedId is removed from the user's banned list
      operationId: unbanUser
      responses:
        "204":
          description: bannedId is removed from the user's banned list
        "400":
          description: bannedId isn't already in the user's banned list
        "401":
          description: the user has no right to access this resource
        "404":
          description: can't find bannedId
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

#######################################################################
#######################################################################
#######################################################################
components:
  schemas:
...
    photo:
      description: |
        an object representing a photo published by someone: it includes an identifier for
        the photo, the username of the user who published the photo, the image (represented
        by its identifier and the path to the folder in which the image is stored), the date in which
        the photo was published, an array containing the partial profile of the users that like the photo,
        an array containing the comments published on the photo, how many likes and comments were left on the photo
      type: object
      properties:
        photoIdentifier:
          $ref: "#/components/schemas/photoId"
        owner:
          $ref: "#/components/schemas/username"
        image:
          $ref: "#/components/schemas/image"
        updateDate:
          description: the image's update date
          type: string
          format: date
          minLength: 10
          maxLength: 10
          example: "2023-12-09"
        likesOnPhoto:
          description: |
            an array containing the username of all the users
            that liked the photo
          type: array
          minItems: 0
          maxItems: 99999
          items:
            $ref: "#/components/schemas/partialProfile"
        likeCount:
          description: how many likes are left on the photo
          type: integer
          minimum: 0
          maximum: 99999
          example: 10
        commentsOnPhoto:
          description: an array containing all the comments published on the photo
          type: array
          minItems: 0
          maxItems: 99999
          items:
            $ref: "#/components/schemas/comment"
        commentCount:
          description: how many comments are published under the photo
          type: integer
          minimum: 0
          maximum: 99999
          example: 5

  ###################################################################
    comment:
      description: |
        an object representing an identifier for the comment, the partial profile
        of the user who wrote the comment, the id of the photo on which he wrote the comment,
        the text that he wrote as a comment and the date in which he wrote it
      type: object
      properties:
        commentIdentifier:
          $ref: "#/components/schemas/commentId"
        userWriter:
          $ref: "#/components/schemas/partialProfile"
        onPhoto:
          $ref: "#/components/schemas/photoId"
        saying:
          $ref: "#/components/schemas/commentOnReq"
        publishingDate:
          description: the date in which the comment was published
          type: string
          format: date
          minLength: 10
          maxLength: 10
          example: "2023-12-09"

    ###################################################################
    partialProfile:
      description: an object representing a user's personal infos (userid, username and profile picture)
      type: object
      properties:
        userid:
          $ref: "#/components/schemas/userid"
        username:
          $ref: "#/components/schemas/username"
        profilePict:
          $ref: "#/components/schemas/image"

    ###################################################################
    commentOnReq:
      description: the string that the user wants to publish as a comment
      type: string
      pattern: '^.*?$'
      minLength: 1
      maxLength: 300
      example: "Nice photo!"

    ###################################################################
...
    username:
      description: the username chosen by the user
      type: string
      example: Maria
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16

  ###################################################################
    userid:
      description: this is the user id returned by the doLogin function
      type: string
      example: "abcdef012345"
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      readOnly: true

      ###################################################################
    followers:
      description: |
        an array of users,
        each one represented by his partial profile
      type: array
      minItems: 0
      maxItems: 99999
      items:
        $ref: "#/components/schemas/partialProfile"

  ###################################################################
    following:
      description: |
        an array of users,
        each one represented by his partial profile
      type: array
      minItems: 0
      maxItems: 99999
      items:
          $ref: "#/components/schemas/partialProfile"

  ###################################################################
    image:
      description: |
        an object representing the image identifier and the path to the folder in which the image is stored
      type: object
      properties:
        imageIdentifier:
          $ref: "#/components/schemas/imageId"
        pathToImage:
          description: the path to the folder in which the image is stored
          type: string
          pattern: '^[/].*?$'
          minLength: 1
          maxLength: 99999
          example: "/tmp/media/pippo/photos/123"

  ###################################################################
    imageId:
      description: the unique identifier for an image
      type: integer
      minimum: 0
      example: 1234
      readOnly: true

  ###################################################################
...
    photoId:
      description: the unique identifier for a photo published by someone
      type: integer
      minimum: 0
      example: 1234
      readOnly: true

  ###################################################################
    commentId:
      description: the unique identifier for a photo published by someone
      type: integer
      minimum: 0
      example: 1234
      readOnly: true

  ###################################################################
  ###################################################################
  ###################################################################
  parameters:
    userid:
      name: uid
      in: path
      description: this is the user id returned by the doLogin function
      required: true
      schema:
        $ref: "#/components/schemas/userid"

  ###################################################################
...
    bannedUserId:
      name: bannedId
      in: path
      description: this is the user id of a user
      required: true
      schema:
        $ref: "#/components/schemas/userid"

  ###################################################################
    photoParam:
      name: photo
      in: path
      required: true
      description: a photo published by someone
      schema:
        $ref: "#/components/schemas/photo"

  ###################################################################
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer

  ###################################################################